Add files via upload
The file vil_pollen.R contains the commands to calculate pollen percentages for each individual taxon and the functional groups, as well as for creating the raw diagram plots. pollen.csv contains the raw pollen counts for the Villarquemado pollen sequence and the dictionary.csv file defines the attributes for the pollen taxa. Both csv files are needed to create the pollen diagrams.